# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  email?: string;
  name?: string;
  role?: string;
  isActive?: boolean;
  password?: string;
  /** @nullable */
  employeeId?: number | null;
}